+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The structure of D3D11GraphicsPlugin's directory was changed. CD3D11Buffer.h/cpp, CD3D11Texture.h/cpp, CD3D11Shader.h/cpp files were replaced with CD3D11Resources.h/cpp.
 - The structure of GLGraphicsContext's directory was changed. COGLBuffer.h/cpp, COGLTexture.h/cpp, COGLShader.h/cpp files were replaced with COGLResources.h/cpp.
 
+- The API of **IGraphicsObjectManager** was refactored and CreateBlendState/CreateRasterizerState/CreateDepthStencilState methods were removed from that.
+
 ### Fixed
 
 - CBaseMaterial::Save's behaviour was fixed.